requirements.txt
Pillow>=9.0.0
requests>=2.28.0
PyQt5>=5.15.0
typing-extensions>=4.0.0
